requirements.txt
PySide6>=6.4.0
Pillow>=9.0.0
